git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@5797 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/USER-REAXC/fix_reax_c.cpp
@@ -152,5 +152,5 @@ void FixReaxC::unpack_comm(int n, int first, double *buf)
   m = 0;
   last = first + n;
   for (i = first; i < last; i++)
-    num_bonds[i] = buf[m++];
+    num_bonds[i] = static_cast<int> (buf[m++]);
 }
